requirements.txt
boto3
slack_sdk
PyYAML
